requirements.txt
flask==3.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-pinecone==0.0.3
openai==1.12.0
pinecone-client==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
